introduced page component for clients
